+++ src/interpolate.cpp
@@ -58,7 +58,7 @@ double interpolate( vector<double> xData, vector<double> yData, double x, bool e
 //======================================================================
 double interpolate( vector<int> xData, vector<double> yData, int x, bool extrapolate )
 {
-   int size = xData.size();
+   unsigned int size = static_cast<unsigned int>(xData.size());
 
    int i = 0;                                                                  // find left end of interval for interpolation
    if ( x >= xData[size - 2] )                                                 // special case: beyond right end
