+++ src/multi_line.cpp
@@ -75,7 +75,7 @@ void CGeomMultiLine::AppendLineSegment(vector<pair<int, int> >* pprVIn)
 //! Returns the CGeomMultiLine object's number of line segments
 int CGeomMultiLine::nGetNumLineSegments(void) const
 {
-   return m_prVVLineSegment.size();
+   return static_cast<int>(m_prVVLineSegment.size());
 }
 
 //! Cuts short the number of line segments
@@ -94,7 +94,7 @@ void CGeomMultiLine::InsertLineSegment(int const nSegment)
 
    // Store the profile numbers that are in this existing vector of pairs, these are the profiles that will be affected by this insertion
    vector<int> nVProfsAffected;
-   for (int i = 0; i < prVPrev.size(); i++)
+   for (unsigned int i = 0; i < prVPrev.size(); i++)
       nVProfsAffected.push_back(prVPrev[i].first);
 
    vector<vector<pair<int, int> > >::iterator it;
@@ -103,11 +103,11 @@ void CGeomMultiLine::InsertLineSegment(int const nSegment)
    m_prVVLineSegment.insert(it+nSegment+1, prVPrev);
 
    // Must now increment the profile's own line seg numbers, but only for those profile numbers which were affected by the insertion. Do this for the new line seg and every line seg after that
-   for (int m = nSegment+1; m < m_prVVLineSegment.size(); m++)
+   for (unsigned int m = nSegment+1; m < m_prVVLineSegment.size(); m++)
    {
-      for (int n = 0; n < m_prVVLineSegment[m].size(); n++)
+      for (unsigned int n = 0; n < m_prVVLineSegment[m].size(); n++)
       {
-         for (int i = 0; i < nVProfsAffected.size(); i++)
+         for (unsigned int i = 0; i < nVProfsAffected.size(); i++)
          {
             if (m_prVVLineSegment[m][n].first == nVProfsAffected[i])
                m_prVVLineSegment[m][n].second++;
@@ -120,7 +120,7 @@ void CGeomMultiLine::InsertLineSegment(int const nSegment)
 vector<vector<pair<int, int> > > CGeomMultiLine::prVVGetAllLineSegAfter(int const nSegment)
 {
    vector<vector<pair<int, int> > > prVTmp;
-   for (int n = nSegment; n < m_prVVLineSegment.size(); n++)
+   for (unsigned int n = nSegment; n < m_prVVLineSegment.size(); n++)
       prVTmp.push_back(m_prVVLineSegment[n]);
 
    return prVTmp;
@@ -136,7 +136,7 @@ void CGeomMultiLine::RemoveLineSegment(int const nSegment)
 //! Appends a coincident profile pair to the CGeomMultiLine object's final line segment
 void CGeomMultiLine::AppendCoincidentProfileToLineSegments(pair<int, int> const prIn)
 {
-   int nSize = m_prVVLineSegment.size();
+   long unsigned int nSize = m_prVVLineSegment.size();
    m_prVVLineSegment[nSize-1].push_back(prIn);
 //   m_prVVLineSegment.back().push_back(prIn);
 }
@@ -166,18 +166,18 @@ int CGeomMultiLine::nGetCoincidentProfileForLineSegment(int const nSegment, int 
 //! Returns the number of coincident profiles in a specified line segment
 int CGeomMultiLine::nGetNumCoincidentProfilesInLineSegment(int const nSegment)
 {
-   return m_prVVLineSegment[nSegment].size();
+   return static_cast<int>(m_prVVLineSegment[nSegment].size());
 }
 
 //! Returns true if the given profile number is amongst the coincident profiles of the CGeomMultiLine object's final line segment
 bool CGeomMultiLine::bFindProfileInCoincidentProfilesOfLastLineSegment(int const nProfile)
 {
-   int nLineSegSize = m_prVVLineSegment.size();
+   long unsigned int nLineSegSize = m_prVVLineSegment.size();
       
    // Note no check to ensure that nLineSegSize < 0
-   int nCoincidentSize = m_prVVLineSegment[nLineSegSize-1].size();
+   long unsigned int nCoincidentSize = m_prVVLineSegment[nLineSegSize-1].size();
 
-   for (int i = 0; i < nCoincidentSize; i++)
+   for (unsigned int i = 0; i < nCoincidentSize; i++)
    {
       if (m_prVVLineSegment[nLineSegSize-1][i].first == nProfile)
          return true;
@@ -202,13 +202,13 @@ bool CGeomMultiLine::bFindProfileInCoincidentProfilesOfLastLineSegment(int const
 //! Returns true if the given profile number is a coincident profile of any line segment of the CGeomMultiLine object
 bool CGeomMultiLine::bFindProfileInCoincidentProfiles(int const nProfile)
 {
-   int nSegSize = m_prVVLineSegment.size();
+   int nSegSize = static_cast<int>(m_prVVLineSegment.size());
    if (nSegSize == 0)
       return false;
 
    for (int i = nSegSize-1; i >= 0; i--)
    {
-      for (int j = 0; j < m_prVVLineSegment[i].size(); j++)
+      for (unsigned int j = 0; j < m_prVVLineSegment[i].size(); j++)
       {
          if (m_prVVLineSegment[i][j].first == nProfile)
          return true;
@@ -224,13 +224,13 @@ void CGeomMultiLine::GetMostCoastwardSharedLineSegment(int const nOtherProfile, 
    nThisLineSegment =
    nOtherLineSegment = -1;
 
-   int nSegSize = m_prVVLineSegment.size();
+   long unsigned int nSegSize = m_prVVLineSegment.size();
    if (nSegSize == 0)
       return;
 
-   for (int i = 0; i < nSegSize; i++)
+   for (unsigned int i = 0; i < nSegSize; i++)
    {
-      for (int j = 0; j < m_prVVLineSegment[i].size(); j++)
+      for (unsigned int j = 0; j < m_prVVLineSegment[i].size(); j++)
       {
          if (m_prVVLineSegment[i][j].first == nOtherProfile)
          {
@@ -270,7 +270,7 @@ int CGeomMultiLine::nFindProfilesLastSeg(int const nProfile) const
    int nSeg = -1;
    for (int i = static_cast<int>(m_prVVLineSegment.size()-1); i >= 0; i--)
    {
-      for (int j = 0; j < m_prVVLineSegment[i].size(); j++)
+      for (unsigned int j = 0; j < m_prVVLineSegment[i].size(); j++)
       {
          if (m_prVVLineSegment[i][j].first == nProfile)
             nSeg = i;
